Separated agent and server logic
