+++ internal/api/server.go
@@ -4,17 +4,11 @@ import (
 	"errors"
 	"log"
 	"net/http"
-	"time"
 
 	"github.com/go-chi/chi/v5"
 
 	"go-yandex-metrics/internal/storage"
 )
-
-type Agent struct {
-	store storage.MemStorage
-	cfg   AgentCfg
-}
 
 type Server struct {
 	store  storage.MemStorage
@@ -53,34 +47,4 @@ func (s *Server) routes() {
 		r.Get("/value/{mtype}/{mname}", s.GetHandler)
 		r.Post("/update/{mtype}/{mname}/{mvalue}", s.UpdateHandler)
 	})
-}
-
-func NewAgent(cfg AgentCfg, store *storage.MemStorage) *Agent {
-	agt := &Agent{
-		cfg:   cfg,
-		store: *store,
-	}
-
-	return agt
-}
-func (a *Agent) Start() error {
-	tickerSave := time.NewTicker(time.Duration(a.cfg.PollInterval) * time.Second)
-	tickerSend := time.NewTicker(time.Duration(a.cfg.ReportInterval) * time.Second)
-
-	for {
-		select {
-		case <-tickerSave.C:
-			err := a.SaveMetrics()
-			if err != nil {
-				log.Println("failed to save metrics: %w", err)
-				return nil
-			}
-		case <-tickerSend.C:
-			err := a.SendMetrics()
-			if err != nil {
-				log.Println("failed to send metrics: %w", err)
-				return nil
-			}
-		}
-	}
 }
